requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9

torch==2.4.1
torchvision==0.19.1

pillow==10.4.0
numpy==1.26.4
pytest==8.3.2
httpx==0.27.0
pyyaml==6.0.2
joblib==1.4.2